Correct consturctor calls #7

+++ tetengo2_detail/include/tetengo2/detail/windows/message_handler_detail/abstra.h
@@ -114,7 +114,7 @@ namespace tetengo2 { namespace detail { namespace windows { namespace message_ha
                 {
                     BOOST_THROW_EXCEPTION((
                         std::system_error{
-                            std::error_code{ ::GetLastError(), win32_category() },
+                            std::error_code{ static_cast<int>(::GetLastError()), win32_category() },
                             "Can't obtain the dropped file path."
                         }
                     ));
