+++ tetengo2_detail/include/tetengo2/detail/windows/message_handler_detail/widget.h
@@ -361,7 +361,8 @@ namespace tetengo2 { namespace detail { namespace windows { namespace message_ha
             {
                 BOOST_THROW_EXCEPTION((
                     std::system_error{
-                        std::error_code{ ::GetLastError(), win32_category() }, "Can't obtain scroll information."
+                        std::error_code{ static_cast<int>(::GetLastError()), win32_category() },
+                        "Can't obtain scroll information."
                     }
                 ));
             }
