+++ tetengo2_detail/include/tetengo2/detail/windows/unit.h
@@ -172,7 +172,8 @@ namespace tetengo2 { namespace detail { namespace windows
             {
                 BOOST_THROW_EXCEPTION((
                     std::system_error{
-                        std::error_code{ ::GetLastError(), win32_category() }, "Can't get the desktop device context."
+                        std::error_code{ static_cast<int>(::GetLastError()), win32_category() },
+                        "Can't get the desktop device context."
                     }
                 ));
             }
